+++ source/Plugins/SymbolFile/DWARF/DWARFASTParserRust.h
@@ -73,7 +73,7 @@ private:
   lldb::TypeSP ParseStructureType(const DWARFDIE &die);
   lldb::TypeSP ParseCLikeEnum(lldb_private::Log *log, const DWARFDIE &die);
 
-  std::vector<unsigned> ParseDiscriminantPath(const char **in_str);
+  std::vector<size_t> ParseDiscriminantPath(const char **in_str);
 
   lldb_private::RustASTContext &m_ast;
